+++ dbldatagen/data_generator.py
@@ -12,7 +12,8 @@ import re
 from pyspark.sql.types import LongType, IntegerType, StringType, StructType, StructField, DataType
 from .spark_singleton import SparkSingleton
 from .column_generation_spec import ColumnGenerationSpec
-from .datagen_constants import DEFAULT_RANDOM_SEED, RANDOM_SEED_FIXED, RANDOM_SEED_HASH_FIELD_NAME, MIN_SPARK_VERSION
+from .datagen_constants import DEFAULT_RANDOM_SEED, RANDOM_SEED_FIXED, RANDOM_SEED_HASH_FIELD_NAME, \
+                               DEFAULT_SEED_COLUMN, SPARK_RANGE_COLUMN, MIN_SPARK_VERSION
 from .utils import ensure, topologicalSort, DataGenError, deprecated
 from . _version import _get_spark_version
 
@@ -28,6 +29,8 @@ _OLD_MAX_OPTION = 'max'
 RATE_SOURCE = "rate"
 RATE_PER_MICRO_BATCH_SOURCE = "rate-micro-batch"
 SPARK_RATE_MICROBATCH_VERSION = "3.2.1"
+
+_STREAMING_TIMESTAMP_COLUMN = "_source_timestamp"
 
 
 class DataGenerator:
@@ -45,6 +48,12 @@ class DataGenerator:
     :param verbose: = if `True`, generate verbose output
     :param batchSize: = UDF batch number of rows to pass via Apache Arrow to Pandas UDFs
     :param debug: = if set to True, output debug level of information
+    :param seedColumnName: = if set, this should be the name of the `seed` or logical `id` column. Defaults to `id`
+
+    By default the seed column is named `id`. If you need to use this column name in your generated data,
+    it is recommended that you use a different name for the seed column - for example `_id`.
+
+    This may be specified by setting the `seedColumnName` attribute to `_id`
     """
 
     # class vars
@@ -62,7 +71,8 @@ class DataGenerator:
 
     def __init__(self, sparkSession=None, name=None, randomSeedMethod=None,
                  rows=1000000, startingId=0, randomSeed=None, partitions=None, verbose=False,
-                 batchSize=None, debug=False, **kwargs):
+                 batchSize=None, debug=False, seedColumnName=DEFAULT_SEED_COLUMN,
+                 **kwargs):
         """ Constructor for data generator object """
 
         # set up logging
@@ -70,6 +80,11 @@ class DataGenerator:
         self.debug = debug
 
         self._setupLogger()
+        self._seedColumnName = seedColumnName
+        self._outputStreamingFields = False
+
+        if seedColumnName != DEFAULT_SEED_COLUMN:
+            self.logger.info(f"Using '{self._seedColumnName}' for seed column in place of '{DEFAULT_SEED_COLUMN}")
 
         self.name = name if name is not None else self.generateName()
         self._rowCount = rows
@@ -139,7 +154,9 @@ class DataGenerator:
         self._buildOrder = []
         self._inferredSchemaFields = []
         self.buildPlanComputed = False
-        self.withColumn(ColumnGenerationSpec.SEED_COLUMN, LongType(), nullable=False, implicit=True, omit=True)
+
+        # lets add the seed column
+        self.withColumn(self._seedColumnName, LongType(), nullable=False, implicit=True, omit=True, noWarn=True)
         self._batchSize = batchSize
 
         # set up spark session
@@ -147,6 +164,11 @@ class DataGenerator:
 
         # set up use of pandas udfs
         self._setupPandas(batchSize)
+
+    @property
+    def seedColumnName(self):
+        """ return the name of data generation seed column"""
+        return self._seedColumnName
 
     @classmethod
     def _checkSparkVersion(cls, sparkVersion, minSparkVersion):
@@ -301,9 +323,12 @@ class DataGenerator:
         if not self.buildPlanComputed:
             self.computeBuildPlan()
 
+        rc = self._rowCount
+        tasks = self.partitions
         output = ["", "Data generation plan", "====================",
-                  f"spec=DateGenerator(name={self.name}, rows={self._rowCount}, startingId={self.starting_id}, partitions={self.partitions})"
-            , ")", "", f"column build order: {self._buildOrder}", "", "build plan:"]
+                  f"spec=DateGenerator(name={self.name}, rows={rc}, startingId={self.starting_id}, partitions={tasks})"
+                  , ")", "", f"seed column: {self._seedColumnName}", "",
+                  f"column build order: {self._buildOrder}", "", "build plan:"]
 
         for plan_action in self._buildPlan:
             output.append(" ==> " + plan_action)
@@ -355,14 +380,14 @@ class DataGenerator:
         return self._rowCount
 
     def withIdOutput(self):
-        """ output seed column field (defaults to `id`) as a column in the test data set if specified
+        """ output seed column field (defaults to `id`) as a column in the generated data set if specified
 
-        If this is not called, the seed column field is omitted from the final test data set
+        If this is not called, the seed column field is omitted from the final generated data set
 
         :returns: modified in-place instance of test data generator allowing for chaining of calls
                   following Builder pattern
         """
-        self._columnSpecsByName[ColumnGenerationSpec.SEED_COLUMN].omit = False
+        self._columnSpecsByName[self._seedColumnName].omit = False
         self._markForPlanRegen()
 
         return self
@@ -594,7 +619,7 @@ class DataGenerator:
 
         all_fields = self.getInferredColumnNames()
         effective_fields = [x for x in all_fields if
-                            (fields is None or x in fields) and x != ColumnGenerationSpec.SEED_COLUMN]
+                            (fields is None or x in fields) and x != self._seedColumnName]
 
         if patterns is not None:
             effective_fields = [x for x in effective_fields for y in patterns if re.search(y, x) is not None]
@@ -615,7 +640,7 @@ class DataGenerator:
         :returns: True if test passes
         """
         inferred_columns = self.getInferredColumnNames()
-        if allowId and columns == ColumnGenerationSpec.SEED_COLUMN:
+        if allowId and columns == self._seedColumnName:
             return True
 
         if type(columns) is list:
@@ -687,14 +712,30 @@ class DataGenerator:
     def withColumn(self, colName, colType=StringType(), minValue=None, maxValue=None, step=1,
                    dataRange=None, prefix=None, random=False, distribution=None,
                    baseColumn=None, nullable=True,
-                   omit=False, implicit=False,
+                   omit=False, implicit=False, noWarn=False,
                    **kwargs):
-        """ add a new column for specification
+        """ add a new column to the synthetic data generation specification
+
+        :param colName: Name of column to add. If this conflicts with the underlying seed column (`id`), it is
+                        recommended that the seed column name is customized during the construction of the data
+                        generator spec.
+        :param colType: Data type for column. This may be specified as either a type from one of the possible
+                        pyspark.sql.types (e.g. `StringType`, `DecimalType(10,3)` etc) or as a string containing a Spark
+                        SQL type definition (i.e  `String`, `array<Integer>`, `map<String, Float>`)
+        :param omit: if True, the column will be omitted from the final set of columns in the generated data.
+                     Used to create columns that are used by other columns as intermediate results.
+                     Defaults to False
+
+        :param expr: Specifies SQL expression used to create column value. If specified, overrides the default rules
+                     for creating column value. Defaults to None
+
+        :param baseColumn: String or list of columns to control order of generation of columns. If not specified,
+                           column is dependent on base seed column (which defaults to `id`)
 
         :returns: modified in-place instance of test data generator allowing for chaining of calls
                   following Builder pattern
 
-        You may also add a variety of options to further control the test data generation process.
+        You may also add a variety of additional options to further control the test data generation process.
         For full list of options, see :doc:`/reference/api/dbldatagen.column_spec_options`.
 
         """
@@ -702,6 +743,10 @@ class DataGenerator:
         ensure(colType is not None, f"Must specify column type for column `{colName}`")
         if baseColumn is not None:
             self._checkColumnOrColumnList(baseColumn, allowId=True)
+
+        if not noWarn and colName == self._seedColumnName:
+            self.logger.warning(f"Adding a new column named '{colName}' overrides seed column '{self._seedColumnName}'")
+            self.logger.warning(f"Use `seedColumName` option on DataGenerator construction for different seed column")
 
         # handle migration of old `min` and `max` options
         if _OLD_MIN_OPTION in kwargs:
@@ -784,6 +829,7 @@ class DataGenerator:
                                            nullable=nullable,
                                            verbose=self.verbose,
                                            debug=self.debug,
+                                           seedColumnName=self._seedColumnName,
                                            **new_props)
 
         self._columnSpecsByName[colName] = column_spec
@@ -822,8 +868,10 @@ class DataGenerator:
                                           end=end_id,
                                           numPartitions=id_partitions)
 
-            if ColumnGenerationSpec.SEED_COLUMN != "id":
-                df1 = df1.withColumnRenamed("id", ColumnGenerationSpec.SEED_COLUMN)
+            # spark.range generates a dataframe with the column `id` so rename it if its not our seed column
+            if SPARK_RANGE_COLUMN != self._seedColumnName:
+                df1 = df1.withColumnRenamed(SPARK_RANGE_COLUMN, self._seedColumnName)
+
         else:
             df1 = self._getStreamingBaseDataFrame(startId, options)
 
@@ -943,7 +991,6 @@ class DataGenerator:
         if age_limit_interval is not None:
             df1 = df1.where(f"""abs(cast(now() as double) - cast(`timestamp` as double )) 
                                   < cast({age_limit_interval} as double)""")
-
         return df1
 
     def _computeColumnBuildOrder(self):
@@ -959,8 +1006,8 @@ class DataGenerator:
 
         :returns: the build ordering
         """
-        dependency_ordering = [(x.name, set(x.dependencies)) if x.name != ColumnGenerationSpec.SEED_COLUMN else (
-            ColumnGenerationSpec.SEED_COLUMN, set())
+        dependency_ordering = [(x.name, set(x.dependencies)) if x.name != self._seedColumnName else (
+            self._seedColumnName, set())
                                for x in self._allColumnSpecs]
 
         # self.pp_list(dependency_ordering, msg="dependencies")
@@ -968,7 +1015,7 @@ class DataGenerator:
         self.logger.info("dependency list: %s", str(dependency_ordering))
 
         self._buildOrder = list(
-            topologicalSort(dependency_ordering, flatten=False, initial_columns=[ColumnGenerationSpec.SEED_COLUMN]))
+            topologicalSort(dependency_ordering, flatten=False, initial_columns=[self._seedColumnName]))
 
         self.logger.info("columnBuildOrder: %s", str(self._buildOrder))
 
@@ -981,7 +1028,7 @@ class DataGenerator:
 
         The build order will be a list of lists - each list specifying columns that can be built at the same time
         """
-        return [x for x in self._buildOrder if x != [ColumnGenerationSpec.SEED_COLUMN]]
+        return [x for x in self._buildOrder if x != [self._seedColumnName]]
 
     def _getColumnDataTypes(self, columns):
         """ Get data types for columns
@@ -1002,7 +1049,7 @@ class DataGenerator:
         self._buildPlan = []
         self.executionHistory = []
         self._processOptions()
-        self._buildPlan.append(f"Build Spark data frame with seed column: {ColumnGenerationSpec.SEED_COLUMN}")
+        self._buildPlan.append(f"Build Spark data frame with seed column: '{self._seedColumnName}'")
 
         # add temporary columns
         for cs in self._allColumnSpecs: